Revert "refactoring for issue #18_CaseStudy2"
