+++ src/argouml-app/src/org/argouml/uml/ui/foundation/core/ActionAddEnumeration.java
@@ -65,10 +65,39 @@ public class ActionAddEnumeration extends AbstractActionNewModelElement {
         Icon icon = ResourceLoaderWrapper.lookupIcon("Enumeration");
         putValue(Action.SMALL_ICON, icon);
     }
-    
-    @Override
-    protected Object createModelElement(Object namespace) {
-        return Model.getCoreFactory().buildEnumeration("", namespace);
-    }
 
+    /*
+     * @see java.awt.event.ActionListener#actionPerformed(java.awt.event.ActionEvent)
+     */
+    public void actionPerformed(ActionEvent e) {
+        Object target = TargetManager.getInstance().getModelTarget();
+        Object ns = null;
+        if (Model.getFacade().isANamespace(target)) {
+            ns = target;
+        }
+        if (Model.getFacade().isAParameter(target)) {
+            if (Model.getFacade().getBehavioralFeature(target) != null) {
+                target = Model.getFacade().getBehavioralFeature(target);
+            }
+        }
+        if (Model.getFacade().isAFeature(target)) {
+            if (Model.getFacade().getOwner(target) != null) {
+                target = Model.getFacade().getOwner(target);
+            }
+        }
+        if (Model.getFacade().isAEvent(target)) {
+            ns = Model.getFacade().getNamespace(target);
+        }
+        if (Model.getFacade().isAClassifier(target)) {
+            ns = Model.getFacade().getNamespace(target);
+        }
+        if (Model.getFacade().isAAssociationEnd(target)) {
+            target = Model.getFacade().getAssociation(target);
+            ns = Model.getFacade().getNamespace(target);
+        }
+        
+        Object newEnum = Model.getCoreFactory().buildEnumeration("", ns);
+        TargetManager.getInstance().setTarget(newEnum);
+        super.actionPerformed(e);
+    }
 }
